MAME 0.92

+++ src/cpu/hd6309/6309dasm.c
@@ -640,7 +640,7 @@ static const char *tfm_s[] = { "%s+,%s+", "%s-,%s-", "%s+,%s", "%s,%s+" };
 static char *hexstring (int address)
 {
 	static char labtemp[10];
-	sprintf (labtemp, "$%04hX", address);
+	sprintf (labtemp, "$%04hX", (unsigned short)address);
 	return labtemp;
 }
